Convert win32-specific code to be cross platform
GLFW already covers the functionality provided by WindowDesc and related
code, so we replace the latter with GLFWwindow and rewrite CreateSurface
to use that.
Cmake now uses system GLFW to build and link against on non-Windows
platforms.
Data paths are corrected to work on case-sensitive file systems.
Added cmake outputs to .gitignore